EdkCompatibilityPkg: allow ConsoleControl protocol to be used
Apple's bootloader requires the ConsoleControl protocol to be
implemented. By adding the Foundation path to the package description,
the ConsoleProtocol can be used in Ovmf without the need of copying it
to Include/Protocol from MdePkg.

The gEfiConsoleControlProtocolGuid is added to the .dec file, to be
consumed by Ovmf.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Reza Jelveh <reza.jelveh@tuhh.de>
Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>