Adds a boot manager option refresh phase for Mac boot.efi loaders.
When refreshing the available boot manager options, this adds an extra pass which scans for block devices with mounted HFS+ volumes. On each of these volumes, if a boot.efi file is found in one of a list of locations, a boot option entry is created for that Mac bootloader. Stale entries are removed.

This fixes the failing installation of macOS High Sierra and later Sierra revisions, as well as in-place point release updates.